+++ js.bolt.surge/flake.nix
@@ -16,13 +16,13 @@
             if pkgs.stdenv.isDarwin then
               pkgs.fetchurl
                 {
-                  url = "https://downloads.slack-edge.com/slack-cli/slack_cli_2.26.0_macOS_64-bit.tar.gz";
-                  sha256 = "0c9dckpr7dm60b5z6zpxrwbs7nxjm02njmf5f3b6arhdf7bqpxp3";
+                  url = "https://downloads.slack-edge.com/slack-cli/slack_cli_2.29.1_macOS_64-bit.tar.gz";
+                  sha256 = "1krzqkv9rsfxyhv4z7c2pa1hgjs73sb7ip2bnc7njl08yapnhavp";
                 }
             else
               pkgs.fetchurl {
-                url = "https://downloads.slack-edge.com/slack-cli/slack_cli_2.26.0_linux_64-bit.tar.gz";
-                sha256 = "1665dpr4ip66hydvvbg7ki4czvp845qic07svhchn8flqk77jcw2";
+                url = "https://downloads.slack-edge.com/slack-cli/slack_cli_2.29.1_linux_64-bit.tar.gz";
+                sha256 = "054i8msn6y6k02dm9930l6nrm3i83axcz6wp4s8izri4kw9gklh3";
               };
           unpackPhase = "tar -xzf $src";
           installPhase = ''
@@ -34,6 +34,7 @@
       {
         devShell = pkgs.mkShell {
           buildInputs = [
+            pkgs.bash
             pkgs.heroku
             pkgs.nodejs_20
             pkgs.opentofu
@@ -41,6 +42,7 @@
           ];
           shellHook = ''
             export SLACK_CONFIG_DIR="$HOME/.config/slack"
+            export SLACK_ENVIRONMENT_TAG="development"
             mkdir -p $SLACK_CONFIG_DIR
           '';
         };