+++ js.bolt.surge/infra/dyno.tf
@@ -2,13 +2,24 @@ resource "heroku_app" "surged" {
   name   = var.heroku_app_name
   region = "us"
   stack  = "heroku-22"
+  acm = "true"
 
   buildpacks = [ "heroku/nodejs" ]
 
   sensitive_config_vars = {
-     SLACK_APP_TOKEN = var.slack_app_token
-     SLACK_BOT_TOKEN = var.slack_bot_token
+     DATABASE_URL = var.database_url
+     SLACK_CLIENT_ID = var.slack_client_id
+     SLACK_CLIENT_SECRET = var.slack_client_secret
+     SLACK_ENVIRONMENT_TAG = var.slack_environment_tag
+     SLACK_LOG_LEVEL = var.slack_log_level
+     SLACK_SIGNING_SECRET = var.slack_signing_secret
+     SLACK_STATE_SECRET = var.slack_state_secret
    }
+}
+
+resource "heroku_addon" "database" {
+  app_id = heroku_app.surged.id
+  plan   = "heroku-postgresql:essential-0"
 }
 
 resource "heroku_build" "project" {
@@ -19,9 +30,19 @@ resource "heroku_build" "project" {
   }
 }
 
-resource "heroku_formation" "labor" {
+resource "heroku_domain" "surgemail" {
+  app_id   = heroku_app.surged.id
+  hostname = "surgem.ai"
+}
+
+moved {
+  from = heroku_formation.labor
+  to   = heroku_formation.inbox
+}
+
+resource "heroku_formation" "inbox" {
   app_id     = heroku_app.surged.id
-  type       = "worker"
+  type       = "web"
   quantity   = 1
   size       = "eco"
   depends_on = [heroku_build.project]