+++ js.bolt.surge/src/errors/index.js
@@ -1,0 +1,6 @@
+import {
+  EnvironmentVariableInvalidError,
+  EnvironmentVariableMissingError,
+} from "./errors.js";
+
+export { EnvironmentVariableInvalidError, EnvironmentVariableMissingError };